refactor(mangastream): put get_lang_code in helper
