+++ src/rust/mangastream/sources/asurascans/src/helper.rs
@@ -1,4 +1,5 @@
-use aidoku::std::{defaults::defaults_get, String};
+use aidoku::std::String;
+use mangastream_template::helper::*;
 
 pub fn get_base_url() -> String {
 	let code = get_lang_code();
@@ -6,18 +7,6 @@ pub fn get_base_url() -> String {
 		"tr" => String::from("https://asurascanstr.com"),
 		_ => String::from("https://asura.gg"),
 	}
-}
-
-pub fn get_lang_code() -> String {
-	let mut code = String::new();
-	if let Ok(languages) = defaults_get("languages") {
-		if let Ok(arr) = languages.as_array() {
-			if let Ok(language) = arr.get(0).as_string() {
-				code = language.read();
-			}
-		}
-	}
-	code
 }
 
 pub fn get_tag_id(tag: String) -> String {
